Show key power_supply:maxcurrent in power_supply icon tooltip.
On mobile devices the tooltip is usually shown if the respective icon is
touched a little bit longer.

The reason to show this as tooltip only is that I think this should better
be mapped as a feature of power supply units mapped on premise as amenity =
power_supply (closes #54).

I will also add a text to be rendered in campsite feature overlay layer in
addition to the plug symbol on these units.